Extract commander to a separate place
