+++ lib/core/src/server/build-static.js
@@ -1,39 +1,16 @@
 import webpack from 'webpack';
-import program from 'commander';
 import path from 'path';
 import fs from 'fs';
-import chalk from 'chalk';
 import shelljs from 'shelljs';
 import { logger } from '@storybook/node-logger';
-import { parseList, getEnvConfig } from './utils';
+import { getProdCli } from './cli';
 import './config/env';
 import loadConfig from './config';
 
 const defaultFavIcon = require.resolve('./public/favicon.ico');
 
-export function buildStatic({ packageJson, ...loadOptions }) {
-  process.env.NODE_ENV = process.env.NODE_ENV || 'production';
-
-  program
-    .version(packageJson.version)
-    .option('-s, --static-dir <dir-names>', 'Directory where to load static files from', parseList)
-    .option('-o, --output-dir [dir-name]', 'Directory where to store built files')
-    .option('-c, --config-dir [dir-name]', 'Directory where to load Storybook configurations from')
-    .option('-w, --watch', 'Enable watch mode')
-    .parse(process.argv);
-
-  logger.info(chalk.bold(`${packageJson.name} v${packageJson.version}\n`));
-
-  // The key is the field created in `program` variable for
-  // each command line argument. Value is the env variable.
-  getEnvConfig(program, {
-    staticDir: 'SBCONFIG_STATIC_DIR',
-    outputDir: 'SBCONFIG_OUTPUT_DIR',
-    configDir: 'SBCONFIG_CONFIG_DIR',
-  });
-
-  const configDir = program.configDir || './.storybook';
-  const outputDir = program.outputDir || './storybook-static';
+export function buildStaticStandalone(options) {
+  const { outputDir, staticDir, watch } = options;
 
   // create output directory if not exists
   shelljs.mkdir('-p', path.resolve(outputDir));
@@ -47,14 +24,14 @@ export function buildStatic({ packageJson, ...loadOptions }) {
   const config = loadConfig({
     configType: 'PRODUCTION',
     corePresets: [require.resolve('./core-preset-prod.js')],
-    configDir,
-    ...loadOptions,
+    ...options,
   });
+
   config.output.path = path.resolve(outputDir);
 
   // copy all static files
-  if (program.staticDir) {
-    program.staticDir.forEach(dir => {
+  if (staticDir) {
+    staticDir.forEach(dir => {
       if (!fs.existsSync(dir)) {
         logger.error(`Error: no such directory to load static files: ${dir}`);
         process.exit(-1);
@@ -77,10 +54,23 @@ export function buildStatic({ packageJson, ...loadOptions }) {
     }
     logger.info('Building storybook completed.');
   };
+
   const compiler = webpack(config);
-  if (program.watch) {
+
+  if (watch) {
     compiler.watch({}, webpackCb);
   } else {
     compiler.run(webpackCb);
   }
+}
+
+export function buildStatic({ packageJson, ...loadOptions }) {
+  const cliOptions = getProdCli(packageJson);
+
+  buildStaticStandalone({
+    ...cliOptions,
+    ...loadOptions,
+    configDir: cliOptions.configDir || './.storybook',
+    outputDir: cliOptions.outputDir || './storybook-static',
+  });
 }
