+++ lib/core/src/server/cli/index.js
@@ -1,0 +1,4 @@
+import getDevCli from './dev';
+import getProdCli from './prod';
+
+export { getDevCli, getProdCli };
