+++ lib/core/src/server/middleware.js
@@ -14,15 +14,15 @@ export const webpackValid = new Promise((resolve, reject) => {
   webpackReject = reject;
 });
 
-export default function(configDir, loadOptions, quiet) {
+export default function(options) {
+  const { configDir } = options;
+
   // Build the webpack configuration using the `getBaseConfig`
   // custom `.babelrc` file and `webpack.config.js` files
   const config = loadConfig({
     configType: 'DEVELOPMENT',
     corePresets: [require.resolve('./core-preset-dev.js')],
-    configDir,
-    quiet,
-    ...loadOptions,
+    ...options,
   });
   const middlewareFn = getMiddleware(configDir);
